crear login y permisos de algunas paginas

+++ base/views.py
@@ -1,9 +1,37 @@
 from django.shortcuts import render, redirect
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib import messages
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            # Redirect to a success page.
+            return redirect('home')
+        else:
+            # Return an 'invalid login' error message.
+            messages.error(request, 'Username or password does not exist')
+
+    return render(request, 'base/login_register.html')
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
 def home(request):
     query = request.GET.get('query') if request.GET.get('query') != None else ''
     rooms = Room.objects.filter(
@@ -24,6 +52,7 @@ def room(request, pk):
     return render(request, 'base/room.html',context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -36,9 +65,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return redirect('home')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -49,8 +83,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return redirect('home')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
